+++ addOns/callhome/CHANGELOG.md
@@ -5,5 +5,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
-- First version.
+- First version supports Check for Updates and News calls.